requirements.txt
asgiref==3.8.1
Django==5.2.2
sqlparse==0.5.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
channels==4.0.0
channels-redis==4.2.0
